Refactor usage record system with factory pattern and background queueing
Update AiUsageRecord services to follow the same pattern as AiAuditLog:
- Add IAiUsageRecordFactory with AiUsageRecordFactory implementation
- Add AiUsageRecordContext for structured context passing
- Add AiUsageRecordResult to encapsulate operation results (similar to AiAuditResponse)
- Add QueueRecordUsageAsync() method using IBackgroundTaskQueue
- Update middleware to use factory and queue method instead of direct persistence
- Move user context capture from middleware to factory
- Improves performance by moving database I/O to background worker

+++ Umbraco.Ai/src/Umbraco.Ai.Core/Analytics/AiUsageAggregationService.cs
@@ -88,9 +88,9 @@ internal sealed class AiUsageAggregationService : IAiUsageAggregationService
                     RequestCount = g.Count(),
                     SuccessCount = g.Count(r => r.Status == "Succeeded"),
                     FailureCount = g.Count(r => r.Status == "Failed"),
-                    InputTokens = g.Sum(r => (long)r.InputTokens),
-                    OutputTokens = g.Sum(r => (long)r.OutputTokens),
-                    TotalTokens = g.Sum(r => (long)r.TotalTokens),
+                    InputTokens = g.Sum(r => r.InputTokens),
+                    OutputTokens = g.Sum(r => r.OutputTokens),
+                    TotalTokens = g.Sum(r => r.TotalTokens),
                     TotalDurationMs = g.Sum(r => r.DurationMs),
                     CreatedAt = DateTime.UtcNow
                 })
